erc20-load-test-tool/script/requirements.txt
py_solc_x==1.1.1
urllib3==1.26.7
web3==5.25.0